Extract a Formatter class
